Convert new Integration files to TypeScript
Per the TypeScript-only rule, the new files added in this PR are now TS:
- Integration.jsx -> Integration.tsx (typed props/state, last-used helper).
- updateProjectApiKey.js -> updateProjectApiKey.ts (typed thunk).
- Update the ProjectApp import to the .tsx file.

generateProjectApiKey was already authored in TypeScript. The existing JS
files only lightly touched by this PR (newProject, RoutingMap, ProjectApp,
SideBar) are left as-is to avoid a churny migration.
